quiz create and update edited question bank added

+++ database/migrations/2023_08_23_181441_quiz_migration.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('title');
             $table->longText('description')->nullable();
             $table->timestamp('finished_at')->nullable();
-            $table->double('duration')->nullable();
-            $table->double('passing_score')->nullable();
+            $table->integer('duration')->nullable();
+            $table->integer('passing_score')->nullable();
             $table->enum('status', ['publish','draft', 'passive'])->default('draft');
 
             $table->timestamps();
